+++ doc/conf.py
@@ -174,11 +174,11 @@ html_theme_options = {}
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ["css"]
+# html_static_path = ["css"]
 
 # These paths are either relative to html_static_path or fully qualified paths
 # (eg. https://...)
-html_css_files = ["custom.css"]
+# html_css_files = ["custom.css"]
 
 # Add any extra paths that contain custom files (such as robots.txt or
 # .htaccess) here, relative to this directory. These files are copied