+++ qpsolvers/solution.py
@@ -86,8 +86,9 @@ class Solution:
     z: Optional[np.ndarray] = None
     z_box: Optional[np.ndarray] = None
 
+    @property
     def found(self) -> bool:
-        """Did the solver find a solution?"""
+        """Check whether the solver found a solution."""
         return self.x is not None
 
     def is_optimal(self, eps_abs: float) -> bool: