+++ tests/test_solve_problem.py
@@ -18,7 +18,7 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with qpsolvers. If not, see <http://www.gnu.org/licenses/>.
 
-"""Tests for the `solve_problem` function."""
+"""Unit tests for the `solve_problem` function."""
 
 import math
 import unittest
@@ -118,7 +118,7 @@ class TestSolveProblem(unittest.TestCase):
                 else 1e-5
                 if solver in ["highs", "osqp"]
                 else 5e-7
-                if solver == "qpswift"
+                if solver in ["clarabel", "qpswift"]
                 else 1e-7
                 if solver == "gurobi"
                 else 1e-8
@@ -334,7 +334,8 @@ for solver in available_solvers:
             f"test_qpsut05_{solver}",
             TestSolveProblem.get_test_qpsut05(solver),
         )
-    if solver not in ["ecos", "qpswift"]:
+    if solver not in ["clarabel", "ecos", "qpswift"]:
+        # clarabel: https://github.com/oxfordcontrol/Clarabel.rs/issues/14
         # See https://github.com/qpsolvers/qpsolvers/issues/159
         # See https://github.com/qpsolvers/qpsolvers/issues/160
         setattr(
@@ -342,7 +343,8 @@ for solver in available_solvers:
             f"test_maros_meszaros_qptest_{solver}",
             TestSolveProblem.get_test_maros_meszaros_qptest(solver),
         )
-    if solver not in ["ecos", "qpswift"]:
+    if solver not in ["clarabel", "ecos", "qpswift"]:
+        # clarabel: https://github.com/oxfordcontrol/Clarabel.rs/issues/14
         # See https://github.com/qpsolvers/qpsolvers/issues/159
         # See https://github.com/qpsolvers/qpsolvers/issues/160
         setattr(
@@ -350,7 +352,8 @@ for solver in available_solvers:
             f"test_infinite_box_bounds_{solver}",
             TestSolveProblem.get_test_infinite_box_bounds(solver),
         )
-    if solver not in ["ecos", "qpswift", "scs"]:
+    if solver not in ["clarabel", "ecos", "qpswift", "scs"]:
+        # clarabel: https://github.com/oxfordcontrol/Clarabel.rs/issues/14
         # See https://github.com/qpsolvers/qpsolvers/issues/159
         # See https://github.com/qpsolvers/qpsolvers/issues/160
         # See https://github.com/qpsolvers/qpsolvers/issues/161